HPSF writing capability added.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/poi/trunk@353321 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hpsf/UnsupportedVariantTypeException.java
@@ -88,7 +88,7 @@ public abstract class UnsupportedVariantTypeException extends HPSFException
     {
         super("HPSF does not yet support the variant type " + variantType + 
               " (" + Variant.getVariantName(variantType) + ", " +
-              HexDump.toHex((int) variantType) + "). If you want support for " +
+              HexDump.toHex(variantType) + "). If you want support for " +
               "this variant type in one of the next POI releases please " +
               "submit a request for enhancement (RFE) to " +
               "<http://nagoya.apache.org/bugzilla/>! Thank you!");
